+++ .idea/deploymentTargetDropDown.xml
@@ -12,6 +12,6 @@
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-06-11T16:34:44.360573700Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-06-15T18:46:47.619615600Z" />
   </component>
 </project>
