+++ app/src/main/res/drawable-anydpi/ic_action_adiciona_aluno.xml
@@ -1,4 +1,5 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:width="24dp"
     android:height="24dp"
     android:viewportWidth="24"
@@ -7,5 +8,6 @@
     android:alpha="0.8">
   <path
       android:fillColor="@android:color/white"
-      android:pathData="M19,13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z"/>
+      android:pathData="M19,13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z"
+      tools:ignore="VectorRaster" />
 </vector>
